# swagger.yaml
openapi: 3.0.0
info:
  title: Jobs Api
  contact: {}
  version: '1.0'
servers:
- url: https://fastidious-naiad-90c0ed.netlify.app/.netlify/functions/index/api/v1
  variables: {}
paths:
  /auth/register:
    post:
      tags:
      - Auth
      summary: registerUser
      description: creates a new user in database
      operationId: registerUser
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/registerUserRequest'
              - example:
                  name: abc
                  email: abc@gmail.com
                  password: abc123
            example:
              name: abc
              email: abc@gmail.com
              password: abc123
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
  /auth/login:
    post:
      tags:
      - Auth
      summary: loginUser
      description: Log in the user
      operationId: loginUser
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/loginUserRequest'
              - example:
                  email: sophiya@gmail.com
                  password: secret
            example:
              email: sophiya@gmail.com
              password: secret
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
  /jobs:
    post:
      tags:
      - Jobs
      summary: createJob
      description: creates a new job
      operationId: createJob
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/createJobRequest'
              - example:
                  company: Servicenow
                  position: Assoc. Software Engineer
                  status: joined
            example:
              company: Servicenow
              position: Assoc. Software Engineer
              status: joined
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
    get:
      tags:
      - Jobs
      summary: getAlljobs
      description: get the list of all the jobs
      operationId: getAlljobs
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
  /jobs/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: The job ID
    get:
      tags:
      - Jobs
      summary: getJob
      description: get the details of all the jobs
      operationId: getJob
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
    delete:
      tags:
      - Jobs
      summary: deleteJob
      description: Delete the particular job
      operationId: deleteJob
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
    patch:
      tags:
      - Jobs
      summary: updateJob
      description: Edit the details of the job
      operationId: updateJob
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/updateJobRequest'
              - example:
                  company: Trilogy
                  position: SDE
                  status: joined
            example:
              company: Trilogy
              position: SDE
              status: joined
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
components:
  schemas:
    registerUserRequest:
      title: registerUserRequest
      required:
      - name
      - email
      - password
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
      example:
        name: abc
        email: abc@gmail.com
        password: abc123
    loginUserRequest:
      title: loginUserRequest
      required:
      - email
      - password
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      example:
        email: sophiya@gmail.com
        password: secret
    createJobRequest:
      title: createJobRequest
      required:
      - company
      - position
      - status
      type: object
      properties:
        company:
          type: string
        position:
          type: string
        status:
          type: string
      example:
        company: Servicenow
        position: Assoc. Software Engineer
        status: joined
    updateJobRequest:
      title: updateJobRequest
      required:
      - company
      - position
      - status
      type: object
      properties:
        company:
          type: string
        position:
          type: string
        status:
          type: string
      example:
        company: Trilogy
        position: SDE
        status: joined
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
tags:
- name: Misc
  description: ''
- name: Auth
  description: ''
- name: Jobs
  description: ''